Scaffolded out keyframe entity

+++ src/audioProject/controller/WaveController.java
@@ -8,20 +8,20 @@ package audioProject.controller;
 public class WaveController {
 
 	private boolean isBeat = false;
-	private double feel = 0.0;
+	private float feel = 0.0f;
 
 	public void setTime(final long newTime) {
 		double s = Math.sin(newTime * .01);
 		isBeat = Math.abs(s) > .99;
 		//System.out.println(s);
-		feel = Math.cos(newTime * .0000001);
+		feel = (float) Math.cos(newTime * .0000001);
 	}
 
 	public boolean isBeat() {
 		return isBeat;
 	}
 
-	public double getFeel() {
+	public float getFeel() {
 		return feel;
 	}
 }
